Use the objects instead of IDs in domain objects
WorkshopRegistration is changed to refer to specific Users and Workshops
instead of their IDs. This should simplify logic when we are looking at
the reservations.

When viewing a users registered workshops, we return the registrations
instead of the Workshops, because we can then show what state the
registration is in.